change Query.Pack code to uint32

+++ join.go
@@ -15,7 +15,7 @@ type Join struct {
 var _ Query = (*Join)(nil)
 
 // Pack implements a part of the Query interface
-func (q *Join) Pack(data *packData, w io.Writer) (byte, error) {
+func (q *Join) Pack(data *packData, w io.Writer) (uint32, error) {
 	enc := msgpack.NewEncoder(w)
 
 	enc.EncodeMapLen(1)
